+++ lib/mod_manager/smol3.dart
@@ -8,7 +8,7 @@ import 'package:flutter_color/flutter_color.dart';
 import 'package:flutter_context_menu/flutter_context_menu.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:pluto_grid/pluto_grid.dart';
+import 'package:pluto_grid_plus/pluto_grid_plus.dart';
 import 'package:trios/dashboard/changelogs.dart';
 import 'package:trios/dashboard/mod_list_basic_entry.dart';
 import 'package:trios/mod_manager/mod_manager_logic.dart';
@@ -272,7 +272,7 @@ class _Smol3State extends ConsumerState<Smol3>
                         style: PlutoGridStyleConfig.dark(
                           enableCellBorderHorizontal: false,
                           enableCellBorderVertical: false,
-                          rowHeight: 40,
+                          // rowHeight: 40,
                           activatedBorderColor: Colors.transparent,
                           inactivatedBorderColor: Colors.transparent,
                           menuBackgroundColor: theme.colorScheme.surface,
@@ -283,7 +283,7 @@ class _Smol3State extends ConsumerState<Smol3>
                           cellColorInEditState: Colors.transparent,
                           cellColorInReadOnlyState: Colors.transparent,
                           gridBorderColor: Colors.transparent,
-                          checkedColor: Colors.transparent,
+                          // cellCheckedColor: Colors.transparent,
                           activatedColor:
                               theme.colorScheme.onSurface.withOpacity(0.1),
                           evenRowColor:
@@ -563,6 +563,7 @@ class _Smol3State extends ConsumerState<Smol3>
         renderer: (rendererContext) => Builder(builder: (context) {
           if (filteredMods.isEmpty) return const SizedBox();
           String? iconPath = rendererContext.cell.value;
+          return SizedBox(height: 100,);
           return iconPath != null
               ? Image.file(
                   iconPath.toFile(),